+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Added support for AWS Secrets Manager
  - Added support for updating the environment variables of a lambda.
  - start, stop, restart for service with `--wait`
+ - Added support for configuring a load balancer for a service.
 
 ## [0.2.41] - 2024-12-06